security: fix critical XSS vulnerability and type inconsistency
CRITICAL SECURITY FIX:
- Add XSS protection to MarkdownRenderer component
- Escape dangerous HTML characters before rendering
- Prevents script injection via malicious markdown content

TYPE CONSISTENCY FIX:
- Change initial_score from int to float in CandidateData
- Ensures consistency with improved_score (float)
- Allows proper decimal score values like 7.5

Addresses critical issues identified in gemini-code-assist[bot] review

+++ mad_spark_multiagent/coordinator.py
@@ -186,7 +186,7 @@ class EvaluatedIdea(TypedDict):
 class CandidateData(TypedDict):
     """Structure for the final data compiled for each candidate idea."""
     idea: str
-    initial_score: int
+    initial_score: float
     initial_critique: str
     advocacy: str
     skepticism: str
